+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - added CHANGELOG.md
 - added "checkbox" column, it can be used as a check list
-- added `Cmd + Shift + R` or `Ctrl + Shift + R` shortcut to clear filters and sort
+- added `Cmd + Shift + E` or `Ctrl + Shift + E` shortcut to clear filters and sort
 - reworked "index" column, move it to config like "checkbox" column
 - reworked "showColumn=none", show "no columns to show" message instead of "index" column only
 - reworked release triggers, now release by version tag